+++ forui/example/lib/sandbox.dart
@@ -23,9 +23,8 @@ class _SandboxState extends State<Sandbox> with SingleTickerProviderStateMixin {
   }
 
   @override
-  Widget build(BuildContext context) => Column(
-        children: [
-        ],
+  Widget build(BuildContext context) => const Column(
+        
       );
 
   @override